+++ atoum/models/category.py
@@ -4,6 +4,8 @@ from django.urls import reverse
 from django.utils import timezone
 from django.utils.html import format_html
 
+from ..utils.text import normalize_text
+
 
 class Category(models.Model):
     """
@@ -95,6 +97,9 @@ class Category(models.Model):
 
     def __str__(self):
         return self.title
+
+    def normalized_title(self):
+        return normalize_text(self.title)
 
     def get_absolute_url(self):
         """