+++ atoum/models/consumable.py
@@ -3,6 +3,8 @@ from django.utils.translation import gettext_lazy as _
 from django.urls import reverse
 from django.utils import timezone
 
+from ..utils.text import normalize_text
+
 
 class Consumable(models.Model):
     """
@@ -55,6 +57,9 @@ class Consumable(models.Model):
 
     def __str__(self):
         return self.title
+
+    def normalized_title(self):
+        return normalize_text(self.title)
 
     def get_assortments(self):
         """