+++ atoum/settings.py
@@ -7,19 +7,19 @@ right after the line which load the default app settings.
 
 """
 
-ASSORTMENT_PAGINATION = 30
+ATOUM_ASSORTMENT_PAGINATION = 30
 """
 Assortment per page limit for pagination, set it to ``None`` to disable
 pagination.
 """
 
-CATEGORY_PAGINATION = 25
+ATOUM_CATEGORY_PAGINATION = 25
 """
 Category per page limit for pagination, set it to ``None`` to disable
 pagination.
 """
 
-PRODUCT_PAGINATION = 50
+ATOUM_PRODUCT_PAGINATION = 50
 """
 Product per page limit for pagination, set it to ``None`` to disable
 pagination.
@@ -29,4 +29,12 @@ ATOUM_BREADCRUMBS_SHOW_HOME = True
 """
 If breadcrumbs should include the site homepage entry or not. This is only working
 in Atoum views.
+"""
+
+ATOUM_INDEXES_DEBUG = False
+"""
+When enabled, the build and update of search indexes will output every rendered content
+for indexed objects.
+
+This can be huge if you have hundreds or more objects to index.
 """